#7 feature : 아이템 이름 필터 검색 추가
