+++ src/main/java/com/darkanddarker/auction/service/search/SearchService.java
@@ -1,5 +1,6 @@
 package com.darkanddarker.auction.service.search;
 
+import com.darkanddarker.auction.dto.search.SearchRequestDto;
 import com.darkanddarker.auction.model.auction.AuctionItem;
 import com.darkanddarker.auction.repository.auction.AuctionItemRepository;
 import com.darkanddarker.auction.service.specification.AuctionItemSpecification;
@@ -16,11 +17,10 @@ import java.util.List;
 @Transactional(readOnly = true)
 public class SearchService {
 
-    private final AuctionItemSpecification auctionItemSpecification;
     private final AuctionItemRepository auctionItemRepository;
 
-    public List<AuctionItem> findItemsBySearchKey(SearchKeyCollection searchKeys) {
-        Specification<AuctionItem> spec = searchKeys.buildDynamicQuery();
+    public List<AuctionItem> findItemsBySearchKey(SearchRequestDto searchRequestDto) {
+        Specification<AuctionItem> spec = searchRequestDto.buildDynamicQuery();
         return auctionItemRepository.findAll(spec);
     }
